Improve Casava 1.8 pair handling.
This improves upon the previous commit that implements Casava 1.8. There is a
serious amount of rework that needs to be done though in general. The threading
utils file isn't able to check read pairs properly either. So abund-filter
which is dependent on that doesn't work either. Better pair checking in the
threading utils file can be done but to it makes more sense to update screed so
it allows you to check paired information.